# [apm] fix service peer service normalization