--- a/pkg/trace/traceutil/normalize.go
+++ b/pkg/trace/traceutil/normalize.go
@@ -65,7 +65,8 @@ func NormalizeService(svc string, lang string) (string, error) {
 		svc = TruncateUTF8(svc, MaxServiceLen)
 		err = ErrTooLong
 	}
-	s := NormalizeTag(svc)
+	// We are normalizing just the tag value.
+	s := NormalizeTagValue(svc)
 	if s == "" {
 		return fallbackService(lang), ErrInvalid
 	}
@@ -83,7 +84,8 @@ func NormalizePeerService(svc string) (string, error) {
 		svc = TruncateUTF8(svc, MaxServiceLen)
 		err = ErrTooLong
 	}
-	s := NormalizeTag(svc)
+	// We are normalizing just the tag value.
+	s := NormalizeTagValue(svc)
 	if s == "" {
 		return "", ErrInvalid
 	}
@@ -113,11 +115,20 @@ func fallbackService(lang string) string {
 
 const maxTagLength = 200
 
-// NormalizeTag applies some normalization to ensure the tags match the backend requirements.
+// NormalizeTag applies some normalization to ensure the full tag_key:tag_value matches the backend requirements.
 func NormalizeTag(v string) string {
+	return normalize(v, true)
+}
+
+// NormalizeTagValue applies some normalization to ensure the tag values match the backend requirements.
+func NormalizeTagValue(v string) string {
+	return normalize(v, false)
+}
+
+func normalize(v string, checkValidStartChar bool) string {
 	// Fast path: Check if the tag is valid and only contains ASCII characters,
 	// if yes return it as-is right away. For most use-cases this reduces CPU usage.
-	if isNormalizedASCIITag(v) {
+	if isNormalizedASCIITag(v, checkValidStartChar) {
 		return v
 	}
 	// the algorithm works by creating a set of cuts marking start and end offsets in v
@@ -169,8 +180,9 @@ func NormalizeTag(v string) string {
 		switch {
 		case unicode.IsLetter(r):
 			chars++
-		case chars == 0:
-			// this character can not start the string, trim
+		// If it's not a unicode letter, and it's the first char, and we're checking the validity of the start char,
+		// we should goto end because the remaining cases are not valid for a start char.
+		case checkValidStartChar && chars == 0:
 			trim = i + jump
 			goto end
 		case unicode.IsDigit(r) || r == '.' || r == '/' || r == '-':
@@ -297,17 +309,21 @@ func normMetricNameParse(name string) (string, bool) {
 	return string(res), true
 }
 
-func isNormalizedASCIITag(tag string) bool {
+func isNormalizedASCIITag(tag string, checkValidStartChar bool) bool {
 	if len(tag) == 0 {
 		return true
 	}
 	if len(tag) > maxTagLength {
 		return false
 	}
-	if !isValidASCIIStartChar(tag[0]) {
-		return false
+	i := 0
+	if checkValidStartChar {
+		if !isValidASCIITagStartChar(tag[0]) {
+			return false
+		}
+		i++
 	}
-	for i := 1; i < len(tag); i++ {
+	for ; i < len(tag); i++ {
 		b := tag[i]
 		// TODO: Attempt to optimize this check using SIMD/vectorization.
 		if isValidASCIITagChar(b) {
@@ -326,10 +342,10 @@ func isNormalizedASCIITag(tag string) bool {
 	return true
 }
 
-func isValidASCIIStartChar(c byte) bool {
+func isValidASCIITagStartChar(c byte) bool {
 	return ('a' <= c && c <= 'z') || c == ':'
 }
 
 func isValidASCIITagChar(c byte) bool {
-	return isValidASCIIStartChar(c) || ('0' <= c && c <= '9') || c == '.' || c == '/' || c == '-'
+	return isValidASCIITagStartChar(c) || ('0' <= c && c <= '9') || c == '.' || c == '/' || c == '-'
 }