--- a/pkg/trace/traceutil/normalize.go
+++ b/pkg/trace/traceutil/normalize.go
@@ -115,20 +115,21 @@ func fallbackService(lang string) string {
 
 const maxTagLength = 200
 
-// NormalizeTag applies some normalization to ensure the full tag_key:tag_value matches the backend requirements.
+// NormalizeTag applies some normalization to ensure the full tag_key:tag_value string matches the backend requirements.
 func NormalizeTag(v string) string {
 	return normalize(v, true)
 }
 
-// NormalizeTagValue applies some normalization to ensure the tag values match the backend requirements.
+// NormalizeTagValue applies some normalization to ensure the tag value matches the backend requirements.
+// It should be used for cases where we have just the tag_value as the input (instead of tag_key:tag_value).
 func NormalizeTagValue(v string) string {
 	return normalize(v, false)
 }
 
-func normalize(v string, checkValidStartChar bool) string {
+func normalize(v string, allowDigitStartChar bool) string {
 	// Fast path: Check if the tag is valid and only contains ASCII characters,
 	// if yes return it as-is right away. For most use-cases this reduces CPU usage.
-	if isNormalizedASCIITag(v, checkValidStartChar) {
+	if isNormalizedASCIITag(v, allowDigitStartChar) {
 		return v
 	}
 	// the algorithm works by creating a set of cuts marking start and end offsets in v
@@ -180,9 +181,9 @@ func normalize(v string, checkValidStartChar bool) string {
 		switch {
 		case unicode.IsLetter(r):
 			chars++
-		// If it's not a unicode letter, and it's the first char, and we're checking the validity of the start char,
+		// If it's not a unicode letter, and it's the first char, and digits are allowed for the start char,
 		// we should goto end because the remaining cases are not valid for a start char.
-		case checkValidStartChar && chars == 0:
+		case allowDigitStartChar && chars == 0:
 			trim = i + jump
 			goto end
 		case unicode.IsDigit(r) || r == '.' || r == '/' || r == '-':