--- a/releasenotes/notes/fix-service-and-peer-service-normalization-d459b8a4c28ae2e4.yaml
+++ b/releasenotes/notes/fix-service-and-peer-service-normalization-d459b8a4c28ae2e4.yaml
@@ -1,0 +1,11 @@
+# Each section from every release note are combined when the
+# CHANGELOG.rst is rendered. So the text needs to be worded so that
+# it does not depend on any information only available in another
+# section. This may mean repeating some details, but each section
+# must be readable independently of the other.
+#
+# Each section note must be formatted as reStructuredText.
+---
+fixes:
+  - |
+    [apm] fix an issue for service and peer.service normalization where names starting with a digit are incorrectly considered as invalid